Update IO processing value ranges; Use floats

+++ src/sys_host.h
@@ -13,18 +13,18 @@ void sys_host_process(struct sp_port* serialport);
 void sys_host_destroy(void);
 
 int sys_host_get_compressor_mode(void);
-int sys_host_get_compressor_release(void);
+float sys_host_get_compressor_release(void);
 int sys_host_get_noisegate_channel(void);
-int sys_host_get_noisegate_decay(void);
-int sys_host_get_noisegate_threshold(void);
-int sys_host_get_pedalboard_gain(void);
+float sys_host_get_noisegate_decay(void);
+float sys_host_get_noisegate_threshold(void);
+float sys_host_get_pedalboard_gain(void);
 
 void sys_host_set_compressor_mode(int mode);
-void sys_host_set_compressor_release(int value);
+void sys_host_set_compressor_release(float value);
 void sys_host_set_noisegate_channel(int channel);
-void sys_host_set_noisegate_decay(int value);
-void sys_host_set_noisegate_threshold(int value);
-void sys_host_set_pedalboard_gain(int value);
+void sys_host_set_noisegate_decay(float value);
+void sys_host_set_noisegate_threshold(float value);
+void sys_host_set_pedalboard_gain(float value);
 
 void sys_host_set_hmi_page(int page);
 void sys_host_set_hmi_subpage(int subpage);
